// src/present/routes/users/types/usersTypes.ts
export type UserIdModel = {
    /**
     * Идентификатор существующего пользователя
     */
    id: string
};

export type UserInputModel = {
    login: string, // Вход; максимальная длина: 10, минимальная длина: 3, шаблон: ^[a-zA-Z0-9_-]*$, должен быть уникальным
    password: string, // Пароль; максимальная длина: 20, минимальная длина: 6
    email: string // Почта; шаблон: ^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$, должна быть уникальной
}; // Входная модель пользователя

export type UserViewModel = {
    id: string, // Идентификатор
    login: string, // Вход; максимальная длина: 10, минимальная длина: 3, шаблон: ^[a-zA-Z0-9_-]*$, должен быть уникальным
    email: string, // Почта; шаблон: ^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$, должна быть уникальной
    createdAt: string // Строка даты создания
}; // Выходная модель пользователя

export enum userFields {
    id = "id", // Идентификатор
    login = "login", // Вход; максимальная длина: 10, минимальная длина: 3, шаблон: ^[a-zA-Z0-9_-]*$, должен быть уникальным
    email = "email", // Почта; шаблон: ^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$, должна быть уникальной
    createdAt = "createdAt" // Дата создания
}; // Возможные поля пользователя

export type TypeUserFields = keyof typeof userFields; // Тип возможных полей пользователя