// src/present/routes/auth/types/authTypes.ts
export type LoginInputModel = {
    loginOrEmail: string, // Вход или почта; должна быть уникальной
    password: string // Пароль; максимальная длина: 20, минимальная длина: 6
}; // Входная модель проверки подлинности пользователя